Move the db related code to its own files.
New files for dbs

New files for dbs in eggnog. Further adjustments.
