+++ q2_moshpit/eggnog/tests/test_method.py
@@ -5,23 +5,17 @@
 #
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
-
 import qiime2
-import os
 import pandas as pd
 import pandas.testing as pdt
-from unittest.mock import patch
 from qiime2.plugin.testing import TestPluginBase
 from q2_types_genomics.feature_data import MAGSequencesDirFmt
 from q2_types_genomics.per_sample_data import ContigSequencesDirFmt
 from q2_types_genomics.genome_data import SeedOrthologDirFmt, OrthologFileFmt
-from q2_types.feature_data import ProteinSequencesDirectoryFormat
 from q2_types_genomics.reference_db import (
-    DiamondDatabaseDirFmt, EggnogRefDirFmt, NCBITaxonomyDirFmt)
-from .._method import (
-  eggnog_diamond_search, eggnog_annotate, fetch_eggnog_db,
-  build_custom_diamond_db
+    DiamondDatabaseDirFmt, EggnogRefDirFmt
 )
+from .._method import eggnog_diamond_search, eggnog_annotate
 
 
 class TestDiamond(TestPluginBase):
@@ -89,85 +83,3 @@ class TestAnnotate(TestPluginBase):
         self.assertEqual(len(objs), 1)
         df = objs[0][1].view(pd.DataFrame)
         pdt.assert_frame_equal(df, exp)
-
-
-class TestBuildDiamondDB(TestPluginBase):
-    package = 'q2_moshpit.eggnog.tests'
-
-    @patch("subprocess.run")
-    def test_build_custom_diamond_db_simple(self, subp_run):
-        # Instantiate input
-        sequences = ProteinSequencesDirectoryFormat()
-
-        # Call function. Patching will make sure nothing is
-        # actually ran
-        diamond_db = build_custom_diamond_db(sequences)
-
-        # Paths to inputs and outputs
-        path_in = os.path.join(str(sequences), "protein-sequences.fasta")
-        path_out = os.path.join(str(diamond_db), "ref_db.dmnd")
-
-        # Check that command was called in the expected way
-        cmd = [
-            "diamond", "makedb",
-            "--verbose",
-            "--in", f"{path_in}",
-            "--db", f"{path_out}",
-            "--threads", "1",
-            '--file-buffer-size', '67108864'
-        ]
-
-        # Check that commands is ran as expected
-        subp_run.assert_called_once_with(cmd, check=True)
-
-    @patch("subprocess.run")
-    def test_build_custom_diamond_db_with_taxonomy(self, subp_run):
-        # Instantiate input
-        sequences = ProteinSequencesDirectoryFormat()
-        taxonomy_data = NCBITaxonomyDirFmt()
-
-        # Call function. Patching will make sure nothing is
-        # actually ran
-        diamond_db = build_custom_diamond_db(sequences, taxonomy_data)
-
-        # Paths to inputs and outputs
-        path_in = os.path.join(str(sequences), "protein-sequences.fasta")
-        path_tax_map = os.path.join(
-            str(taxonomy_data), "prot.accession2taxid.gz"
-            )
-        path_tax_nodes = os.path.join(str(taxonomy_data), "nodes.dmp")
-        path_tax_names = os.path.join(str(taxonomy_data), "names.dmp")
-        path_out = os.path.join(str(diamond_db), "ref_db.dmnd")
-
-        # Check that command was called in the expected way
-        cmd = [
-            "diamond", "makedb",
-            "--verbose",
-            "--in", f"{path_in}",
-            "--db", f"{path_out}",
-            "--threads", "1",
-            '--file-buffer-size', '67108864',
-            "--taxonmap", f"{path_tax_map}",
-            "--taxonnodes", f"{path_tax_nodes}",
-            "--taxonnames", f"{path_tax_names}",
-        ]
-
-        # Check that commands is ran as expected
-        subp_run.assert_called_once_with(cmd, check=True)
-
-
-class TestFetchDB(TestPluginBase):
-    package = 'q2_moshpit.eggnog.tests'
-
-    @patch("subprocess.run")
-    def test_fetch_eggnog_db(self, subp_run):
-        # Call function. Patching will make sure nothing is
-        # actually ran
-        eggnog_db = fetch_eggnog_db()
-
-        # Check that command was called in the expected way
-        cmd = [
-            "download_eggnog_data.py", "-y", "-D",
-            "--data_dir", str(eggnog_db)
-        ]
-        subp_run.assert_called_once_with(cmd, check=True)
